v1.0.14
Added duplicate prevention for bot developer only, fixed autokick not working.
Apparently v1.0.13a-c were not committed to the public build. The gist of it is that people who have no verified yet are now autokicked after being on the server for 1 week, and then 3 discrete bugfix patches for that one feature b/c I'm dummy